requirements.txt
asgiref==3.3.4
Django==3.2.3
psutil==5.8.0
pytz==2021.1
sqlparse==0.4.1

py-cpuinfo~=8.0.0
netaddr~=0.8.0